Update README.md
`sudo npm` is arguably a bad practice. IMHO it shouldn't be recommended. While I've seen some students "need" it (meaning they initially got themselves into a bind, used `sudo`, and must continue to use it), if they "need" `sudo`, their terminal will tell them and they'll figure it out.

Class methods, in my experience, are static methods. They are complementary to member methods, or what we'd call prototype methods in ES5. I've updated the wording to try to be accurate but still be clear to programmers new to React.